+++ packages/ai/src/generate-text/stream-text.test.ts
@@ -12694,7 +12694,9 @@ describe('streamText', () => {
               "message": "Invalid input: expected string, received undefined"
             }
           ]],
-              "input": "{ "cities": "San Francisco" }",
+              "input": {
+                "cities": "San Francisco",
+              },
               "invalid": true,
               "toolCallId": "call-1",
               "toolName": "cityAttractions",
@@ -12713,7 +12715,9 @@ describe('streamText', () => {
               "message": "Invalid input: expected string, received undefined"
             }
           ]",
-              "input": "{ "cities": "San Francisco" }",
+              "input": {
+                "cities": "San Francisco",
+              },
               "toolCallId": "call-1",
               "toolName": "cityAttractions",
               "type": "tool-error",
@@ -12762,7 +12766,9 @@ describe('streamText', () => {
                 "message": "Invalid input: expected string, received undefined"
               }
             ]],
-                "input": "{ "cities": "San Francisco" }",
+                "input": {
+                  "cities": "San Francisco",
+                },
                 "invalid": true,
                 "toolCallId": "call-1",
                 "toolName": "cityAttractions",
@@ -12781,7 +12787,9 @@ describe('streamText', () => {
                 "message": "Invalid input: expected string, received undefined"
               }
             ]",
-                "input": "{ "cities": "San Francisco" }",
+                "input": {
+                  "cities": "San Francisco",
+                },
                 "toolCallId": "call-1",
                 "toolName": "cityAttractions",
                 "type": "tool-error",
@@ -12851,7 +12859,9 @@ describe('streamText', () => {
                 "message": "Invalid input: expected string, received undefined"
               }
             ]",
-                "input": "{ "cities": "San Francisco" }",
+                "input": {
+                  "cities": "San Francisco",
+                },
                 "toolCallId": "call-1",
                 "toolName": "cityAttractions",
                 "type": "tool-input-error",