+++ packages/ai/src/types/index.ts
@@ -1,3 +1,4 @@
+export type { JSONSchema7 } from '@ai-sdk/provider';
 export type { Embedding, EmbeddingModel } from './embedding-model';
 export type {
   ImageModel,
@@ -6,13 +7,13 @@ export type {
 } from './image-model';
 export type { ImageModelResponseMetadata } from './image-model-response-metadata';
 export type { JSONValue } from './json-value';
-export type { JSONSchema7 } from '@ai-sdk/provider';
 export type {
   CallWarning,
   FinishReason,
   LanguageModel,
   ToolChoice,
 } from './language-model';
+export type { LanguageModelMiddleware } from './language-model-middleware';
 export type { LanguageModelRequestMetadata } from './language-model-request-metadata';
 export type { LanguageModelResponseMetadata } from './language-model-response-metadata';
 export type { Provider } from './provider';