+++ packages/elevenlabs/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @ai-sdk/elevenlabs
 
+## 1.0.5
+
+### Patch Changes
+
+- 6aaf688: feat(speech): add elevenlabs text-to-speech
+
+## 1.0.4
+
+### Patch Changes
+
+- Updated dependencies [68751f9]
+  - @ai-sdk/provider-utils@3.0.4
+
 ## 1.0.3
 
 ### Patch Changes